Add files via upload
K Nearest Neighbor Algorithm works with this data in theory - but using this algorithm is not useful at all, since we only have the so called "normal" driving behaviour, and not different categories. 

Look off this for reference if you want - Not sure of the algorithm of best choice yet, looking at that now.